grafana-dashboard-shell-nodes: Fix loading variables
For some reason, Grafana's "load variables on time range change" option
*only* works when the time range is changed... and not when the
dashboard is loaded. I think this should fix it, as this variable
shouldn't change while the dashboard is loaded and it may actually be
preferable to not reload the variables and have it probably change to a
default "first" variable, if it detects that the variables are
different.